add useWatchBalance hook (#829)
